Change float to int for warehouse inventory Ecom orders
git-tfs-id: [http://mlwweb05:8080/tfs/defaultcollection]$/Footlocker.Logistics/Allocation/Main;C34486

+++ Allocation/AllocationLibrary/Models/InventoryReductionsByType.cs
@@ -29,6 +29,6 @@ namespace Footlocker.Logistics.Allocation.Models
         public int RDQQuantity { get; set; }
 
         [Column("OrderQty")]
-        public decimal OrderQuantity { get; set; }
+        public int OrderQuantity { get; set; }
     }
 }
